Set encoding correctly on strings.
Randomization found another bug. :)

The bug was that the Basic Latin characters had encoding US-ASCII
because 32.chr returns " " with an encoding of US-ASCII, regardless
of the environment. The fix was to force encoding. This bug only
showed up when the random text happened to include *only* Basic Latin
characters (usually short text, given probabilities).

I kind of love that this testing randomization library is helping
me find bugs in itself.