Major refactor of CSR and certificate creation
As I read RFC5280 and explore existing SSL certificate usage, the
right abstraction for this certificate stuff still feels a bit unclear.
This iteration feels closer in that it doesn't isolate
root/intermediate/server certificates as crisply, and it reduced a bunch
of code duplication.

Major wiping of the specs, though, since the API has changed quite a
bit. Needs more tests.